requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-redis==6.0.0
gunicorn==25.1.0
psycopg2-binary==2.9.11
whitenoise==6.11.0
python-dotenv==1.2.1

